chore: Completed useSignAndCertify mutation
